requirements.txt
tweepy
pandas
tzlocal==2.1
autocorrect
emoji
dnspython
pymongo[tls,srv]
flask
pymongo
gunicorn
